♻️ [refactor] #140 : 최근 3대 챌린지 기록 시간 구현 옮김
최근 3대 챌린지 기록 시간 구현을 USER API에서 RECORD API로 옮김
