+++ server/src/sbd_records/sbd_records.controller.ts
@@ -30,4 +30,17 @@ export class SbdRecordsController {
   getBestSBDRecord(@Query("userId") userId: number) {
     return this.recordsService.findBestSBDRecord(userId);
   }
+
+  @Get("recent")
+  @ApiOperation({
+    summary:
+      "해당 사용자의 가장 최근 3대 챌린지 기록 시간을 초단위(new Date().getTime() / 1000) 형식으로 반환",
+  })
+  @ApiQuery({
+    name: "id",
+    type: "number",
+  })
+  async getRecentRecordTime(@Query("id") userId: number) {
+    return this.recordsService.getRecentRecordTime(userId);
+  }
 }
