+++ server/src/users/users.service.ts
@@ -62,18 +62,4 @@ export class UsersService {
 
     return user;
   }
-
-  async getRecentRecordTime(userId: number) {
-    // new Date().getTime() / 1000;
-    const record = await this.recordsRepository
-      .createQueryBuilder("record")
-      .where("record.user_id = :userId", { userId })
-      .select("record.time_stamp")
-      .orderBy("record.time_stamp", "DESC")
-      .getRawMany();
-
-    const recentRecord = record[0];
-
-    return { recentRecord };
-  }
 }
